feat(homepage): adding editions
