+++ i18n/en/code.json
@@ -261,6 +261,33 @@
   "home_page.categories": {
     "message": "Categories"
   },
+  "home_page.select_platform": {
+    "message": "Select platform"
+  },
+  "home_page.select_platform_2": {
+    "message": "Get documentation based on the AI Unlimited platform."
+  },
+  "home_page.edition_cloud": {
+    "message": "Cloud service provider"
+  },
+  "home_page.edition_ms_fabric": {
+    "message": "Microsoft Fabric"
+  },
+  "home_page.description.get_started":{
+    "message": "Find the AI Unlimited Workload in Fabric, and prepare a notebook for trying your first use case."
+  },
+  "home_page.ms_fabric_use_case": {
+    "message": "Try the use case"
+  },
+  "home_page.description.use_cases":{
+    "message": "Try the data preparation use case to start. Later you can try the financial customer journey and user segmentation use cases."
+  },
+  "home_page.ms_fabric_analytic_function": {
+    "message": "Analytic functions"
+  },
+  "home_page.description.analytic_function":{
+    "message": "Review our full list of in-database analytics with links to resources on how to use them."
+  },
   "home_page.featured_event": {
     "message": "Featured event"
   },
@@ -569,5 +596,8 @@
   },
   "theme.docs.breadcrumbs.home": {
     "message": "Home page"
+  },
+  "fabric.subjeting": {
+    "message": "Microsoft Fabric"
   }
 }
